+++ src/app/app.module.ts
@@ -1,5 +1,6 @@
 import { BrowserModule } from '@angular/platform-browser';
 import { NgModule } from '@angular/core';
+import { HttpClientModule } from '@angular/common/http';
 
 import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
@@ -9,10 +10,21 @@ import { SharedModule } from './shared/shared.module';
 import { FirstComponent } from './first/components/first/first.component';
 import { CoreModule } from './core/core.module';
 import { LayoutModule } from './layout/layout.module';
+import { httpInterceptorProviders } from './core/interceptors';
 
 @NgModule({
     declarations: [AppComponent, FirstComponent],
-    imports: [BrowserModule, CartModule, ProductsModule, SharedModule, CoreModule, LayoutModule, AppRoutingModule],
+    imports: [
+        BrowserModule,
+        CartModule,
+        ProductsModule,
+        SharedModule,
+        CoreModule,
+        LayoutModule,
+        HttpClientModule,
+        AppRoutingModule,
+    ],
+    providers: [httpInterceptorProviders],
     bootstrap: [AppComponent],
 })
 export class AppModule {}