+++ src/app/products/components/product-view/product-view.component.ts
@@ -1,9 +1,10 @@
 import { Component, OnInit } from '@angular/core';
 import { ActivatedRoute, Router } from '@angular/router';
 
-import { CartService } from '../../../cart/services/cart.service';
+import { CartObservableService } from '../../../cart/services/cart-observable.service';
+import { ICartProductItem } from '../../../cart/models/cart-product.model';
 import { IProductItem } from '../../models/product.model';
-import { ProductsService } from '../../services/products.service';
+import { ProductsPromiseService } from '../../services/products-promise.service';
 
 @Component({
     selector: 'app-product-view',
@@ -11,23 +12,41 @@ import { ProductsService } from '../../services/products.service';
     styleUrls: ['./product-view.component.css'],
 })
 export class ProductViewComponent implements OnInit {
-    product: IProductItem | null;
+    product: IProductItem;
+    cartProducts: ICartProductItem[];
 
     constructor(
         private route: ActivatedRoute,
         private router: Router,
-        public cartService: CartService,
-        public productsService: ProductsService,
+        public productsPromiseService: ProductsPromiseService,
+        public cartObservableService: CartObservableService,
     ) {}
 
     ngOnInit(): void {
-        this.product = this.productsService.getProduct(+this.route.snapshot.params.id);
+        this.productsPromiseService
+            .getProduct(+this.route.snapshot.params.id)
+            .then((product) => (this.product = product))
+            .catch((err) => console.log(err));
+        this.cartObservableService.getCartProducts().subscribe((cartProducts) => {
+            this.cartProducts = cartProducts;
+        });
     }
 
     onBuy(product: IProductItem): void {
-        this.cartService.addCartProduct(product);
+        const itemIndex = this.cartProducts.findIndex((item) => item.id === product.id);
+        if (itemIndex !== -1) {
+            const cartProduct = this.cartProducts[itemIndex];
+            this.cartObservableService
+                .updateCartProduct({ ...cartProduct, quantity: cartProduct.quantity + 1 })
+                .subscribe((cartProducts) => {
+                    this.cartProducts = cartProducts;
+                });
+        } else {
+            this.cartObservableService.createCartProduct(product).subscribe((cartProducts) => {
+                this.cartProducts = cartProducts;
+            });
+        }
     }
-
     onBack(): void {
         this.router.navigate(['/products-list']);
     }